pyproject.toml
[project]
name = "woocommerce-playwright-python"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "Rodel Calasagsag", email = "od.rodel.calasagsag@gmail.com" },
]
dependencies = []
requires-python = "==3.13.*"
readme = "README.md"
license = { text = "MIT" }

[dependency-groups]
test = [
    "pytest-playwright>=0.7.0",
    "python-dotenv>=1.1.0",
    "pytest-pylint>=0.21.0",
    "black>=25.1.0",
]

[tool.pytest.ini_options]
addopts = "--browser chromium --base-url http://localhost:8080 --tracing retain-on-failure --video retain-on-failure --screenshot only-on-failure --full-page-screenshot"
testpaths = ["tests"]

[tool.pdm]
distribution = false

[tool.pdm.scripts]
start = "bash -c 'docker-compose up -d && bash ./store-setup.sh && playwright install chromium'"
stop = "docker-compose down --volumes"
restart = { composite = ["stop", "start"] }
test = "pytest -v"
debug = "env PWDEBUG=1 pytest"
trace = "playwright show-trace"
lint = "py.test --pylint"
format = "black ."
